refactor(#281): split McpPanel.tsx into sub-components
Extract ServerCard, ServerFormFields, PresetList, AddServerForm, and
shared types/constants into apps/desktop/src/components/settings/mcp/.
McpPanel.tsx reduced from 713 to 218 lines.
